Fix:firing queued messages when in other chat
